Feat: Integrate @tanstack/react-query for enhanced data management
- Added @tanstack/react-query to manage server state across the application.
- Updated layout components to include QueryClientProvider for query management.
- Refactored import paths for hooks and services to improve clarity and organization.
- Introduced new hooks for user and chat functionalities to streamline data handling.
