+++ app/(tabs)/_layout.tsx
@@ -6,7 +6,7 @@ import { HapticTab } from '@/shared/components/HapticTab';
 import { IconSymbol } from '@/shared/components/ui/IconSymbol';
 import TabBarBackground from '@/shared/components/ui/TabBarBackground';
 import { Colors } from '@/shared/constants/Colors';
-import { useColorScheme } from '@/shared/hooks/useColorScheme';
+import { useColorScheme } from '@/shared/hooks/theme/useColorScheme';
 
 export default function TabLayout() {
   const colorScheme = useColorScheme();
