+++ package.json
@@ -18,6 +18,7 @@
     "@expo/vector-icons": "^14.0.2",
     "@react-navigation/bottom-tabs": "^7.2.0",
     "@react-navigation/native": "^7.0.14",
+    "@tanstack/react-query": "^5.90.5",
     "babel-plugin-inline-import": "^3.0.0",
     "drizzle-kit": "^0.30.5",
     "drizzle-orm": "^0.40.0",
