+++ shared/components/ThemedView.tsx
@@ -1,6 +1,6 @@
 import { View, type ViewProps } from 'react-native';
 
-import { useThemeColor } from '@/shared/hooks/useThemeColor';
+import { useThemeColor } from '@/shared/hooks/theme/useThemeColor';
 
 export type ThemedViewProps = ViewProps & {
   lightColor?: string;
