+++ shared/hooks/auth/useProtectedRoute.tsx
