+++ shared/hooks/db/useDatabase.ts
@@ -1,0 +1,5 @@
+import { useDatabaseStatus } from "../../database/context/DatabaseProvider";
+
+export function useDatabase() {
+  return useDatabaseStatus();
+}
