--- shared/hooks/useDatabase.ts
@@ -1,5 +1,0 @@
-import { useDatabaseStatus } from "../database/context/DatabaseProvider";
-
-export function useDatabase() {
-  return useDatabaseStatus();
-}
